Move testutils into pkg directory.
