+++ cmds/dd/dd_test.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/u-root/u-root/shared/testutil"
+	"github.com/u-root/u-root/pkg/testutil"
 )
 
 var tests = []struct {
