+++ cmds/dhclient/dhclient_test.go
@@ -10,7 +10,7 @@ import (
 	"os/exec"
 	"testing"
 
-	"github.com/u-root/u-root/shared/testutil"
+	"github.com/u-root/u-root/pkg/testutil"
 )
 
 var tests = []struct {
